+++ haystack/tools/tool_types.py
